archive/requirements.txt
Flask
psycopg2-binary
flask-sqlalchemy
google-cloud-secret-manager
